Breaking change for `agree_ephemeral`
This commit breaks the public API of `agree_ephemeral` by reverting to
four arguments, using an unused error value.

This change is needed to satisfy the expected signature by an older
version of rustls, but while still being compilable to wasm32-wasi.

Again, this should *not* be used in a real world scenario.

Signed-off-by: Radu M <root@radu.sh>